+++ unikmer/cmd/diff.go
@@ -66,6 +66,11 @@ Tips:
 
 		checkFileSuffix(extDataFile, files...)
 
+		var nfiles = len(files)
+		if nfiles <= 1 {
+			checkError(fmt.Errorf("two or more files needed"))
+		}
+
 		outFile := getFlagString(cmd, "out-prefix")
 		sortKmers := getFlagBool(cmd, "sort")
 
@@ -73,17 +78,17 @@ Tips:
 
 		runtime.GOMAXPROCS(threads)
 
-		m := make(map[uint64]struct{}, mapInitSize)
+		m := make(map[uint64]uint32, mapInitSize)
 
 		var infh *bufio.Reader
 		var r *os.File
 		var reader *unikmer.Reader
 		var code uint64
+		var taxid uint32
 		var k int = -1
 		var canonical bool
 		var hasTaxid bool
 		var ok bool
-		var nfiles = len(files)
 
 		// -----------------------------------------------------------------------
 
@@ -91,113 +96,6 @@ Tips:
 		if opt.Verbose {
 			log.Infof("processing file (%d/%d): %s", 1, nfiles, file)
 		}
-
-		// only one file given
-		if len(files) == 1 {
-			func() {
-				infh, r, _, err = inStream(file)
-				checkError(err)
-				defer r.Close()
-
-				reader, err = unikmer.NewReader(infh)
-				checkError(err)
-
-				k = reader.K
-
-				if !isStdout(outFile) {
-					outFile += extDataFile
-				}
-				outfh, gw, w, err := outStream(outFile, opt.Compress, opt.CompressionLevel)
-				checkError(err)
-				defer func() {
-					outfh.Flush()
-					if gw != nil {
-						gw.Close()
-					}
-					w.Close()
-				}()
-
-				var writer *unikmer.Writer
-				var m2 []uint64
-
-				if sortKmers {
-					m2 = make([]uint64, 0, mapInitSize)
-				} else {
-					var mode uint32
-					if sortKmers {
-						mode |= unikmer.UNIK_SORTED
-					} else if opt.Compact {
-						mode |= unikmer.UNIK_COMPACT
-					}
-					if reader.IsCanonical() {
-						mode |= unikmer.UNIK_CANONICAL
-					}
-					if reader.IsIncludeTaxid() {
-						mode |= unikmer.UNIK_INCLUDETAXID
-					}
-					writer, err = unikmer.NewWriter(outfh, reader.K, mode)
-					checkError(err)
-				}
-
-				m := make(map[uint64]struct{}, mapInitSize)
-				for {
-					code, err = reader.ReadCode()
-					if err != nil {
-						if err == io.EOF {
-							break
-						}
-						checkError(err)
-					}
-
-					if _, ok = m[code]; !ok {
-						m[code] = struct{}{}
-						if sortKmers {
-							m2 = append(m2, code)
-						} else {
-							writer.WriteCode(code) // not need to check er
-						}
-					}
-				}
-
-				if sortKmers {
-					var mode uint32
-					if reader.IsCanonical() {
-						mode |= unikmer.UNIK_CANONICAL
-					}
-					if reader.IsIncludeTaxid() {
-						mode |= unikmer.UNIK_INCLUDETAXID
-					}
-					mode |= unikmer.UNIK_SORTED
-					writer, err = unikmer.NewWriter(outfh, reader.K, mode)
-					checkError(err)
-
-					writer.Number = int64(len(m2))
-
-					if opt.Verbose {
-						log.Infof("sorting %d k-mers", len(m2))
-					}
-					sort.Sort(unikmer.CodeSlice(m2))
-					if opt.Verbose {
-						log.Infof("done sorting")
-					}
-
-					for _, code = range m2 {
-						writer.WriteCode(code)
-					}
-				}
-
-				checkError(writer.Flush())
-				if opt.Verbose {
-					log.Infof("%d k-mers saved to %s", len(m), outFile)
-				}
-			}()
-
-			return
-		}
-
-		// -----------------------------------------------------------------------
-
-		// > 1 files
 
 		// read firstFile
 
@@ -212,10 +110,10 @@ Tips:
 		hasTaxid = !opt.IgnoreTaxid && reader.HasTaxidInfo()
 
 		var minCode uint64 = ^uint64(0)
-		if reader.Flag&unikmer.UNIK_SORTED > 0 { // query is sorted
+		if reader.IsSorted() { // query is sorted
 			once := true
 			for {
-				code, err = reader.ReadCode()
+				code, taxid, err = reader.ReadCodeWithTaxid()
 				if err != nil {
 					if err == io.EOF {
 						break
@@ -227,11 +125,11 @@ Tips:
 					minCode = code
 					once = false
 				}
-				m[code] = struct{}{}
+				m[code] = taxid
 			}
 		} else {
 			for {
-				code, err = reader.ReadCode()
+				code, taxid, err = reader.ReadCodeWithTaxid()
 				if err != nil {
 					if err == io.EOF {
 						break
@@ -242,7 +140,7 @@ Tips:
 				if code < minCode {
 					minCode = code
 				}
-				m[code] = struct{}{}
+				m[code] = taxid
 			}
 		}
 
@@ -322,7 +220,7 @@ Tips:
 		chFile := make(chan iFile, threads)
 		doneSendFile := make(chan int)
 
-		maps := make(map[int]map[uint64]struct{}, threads)
+		maps := make(map[int]map[uint64]uint32, threads)
 		maps[0] = m
 
 		if len(files) > 2 {
@@ -333,7 +231,7 @@ Tips:
 			var wg sync.WaitGroup
 			type iMap struct {
 				i int
-				m map[uint64]struct{}
+				m map[uint64]uint32
 			}
 			ch := make(chan iMap, threads)
 			doneClone := make(chan int)
@@ -346,9 +244,9 @@ Tips:
 			for i := 1; i < threads; i++ {
 				wg.Add(1)
 				go func(i int) {
-					m1 := make(map[uint64]struct{}, len(m))
-					for k := range m {
-						m1[k] = struct{}{}
+					m1 := make(map[uint64]uint32, len(m))
+					for k, t := range m {
+						m1[k] = t
 					}
 					ch <- iMap{i: i, m: m1}
 					wg.Done()
@@ -422,16 +320,16 @@ Tips:
 					if reader.IsCanonical() != canonical {
 						checkError(fmt.Errorf(`'canonical' flags not consistent, please check with "unikmer stats"`))
 					}
-					if !opt.IgnoreTaxid && reader.HasTaxidInfo() != hasTaxid {
-						checkError(fmt.Errorf(`taxid information found in some files but missing in others, please check with "unikmer stats"`))
-					}
+					// if !opt.IgnoreTaxid && reader.HasTaxidInfo() != hasTaxid {
+					// 	checkError(fmt.Errorf(`taxid information found in some files but missing in others, please check with "unikmer stats"`))
+					// }
 
 					// file is sorted, so we can skip codes that are small than minCode
-					sorted = reader.Flag&unikmer.UNIK_SORTED > 0
+					sorted = reader.IsSorted()
 					nSkip = 0
 					checkSkip = sorted
 					for {
-						code, err = reader.ReadCode()
+						code, _, err = reader.ReadCodeWithTaxid()
 						if err != nil {
 							if err == io.EOF {
 								break
@@ -496,7 +394,7 @@ Tips:
 		toStop <- 1
 		<-doneDone
 
-		var m0 map[uint64]struct{}
+		var m0 map[uint64]uint32
 		if !hasDiff {
 			if opt.Verbose {
 				log.Infof("no set difference found")
@@ -582,25 +480,25 @@ Tips:
 			checkError(writer.WriteHeader())
 		} else {
 			if sortKmers {
-				codes := make([]uint64, len(m0))
+				codes := make([]unikmer.CodeTaxid, len(m0))
 				i := 0
-				for code := range m0 {
-					codes[i] = code
+				for code, taxid := range m0 {
+					codes[i] = unikmer.CodeTaxid{Code: code, Taxid: taxid}
 					i++
 				}
 				if opt.Verbose {
 					log.Infof("sorting %d k-mers", len(codes))
 				}
-				sort.Sort(unikmer.CodeSlice(codes))
+				sort.Sort(unikmer.CodeTaxidSlice(codes))
 				if opt.Verbose {
 					log.Infof("done sorting")
 				}
-				for _, code := range codes {
-					writer.WriteCode(code)
+				for _, codeT := range codes {
+					writer.WriteCodeWithTaxid(codeT.Code, codeT.Taxid)
 				}
 			} else {
-				for code := range m0 {
-					writer.WriteCode(code)
+				for code, taxid := range m0 {
+					writer.WriteCodeWithTaxid(code, taxid)
 				}
 			}
 		}